+++ Library/Homebrew/compat/requirements/tex_requirement.rb
@@ -1,0 +1,20 @@
+require "requirement"
+
+class TeXRequirement < Requirement
+  fatal true
+  cask "mactex"
+  download "https://www.tug.org/mactex/"
+
+  satisfy { which("tex") || which("latex") }
+
+  def message
+    s = <<~EOS
+      A LaTeX distribution is required for Homebrew to install this formula.
+
+      Make sure that "/usr/texbin", or the location you installed it to, is in
+      your PATH before proceeding.
+    EOS
+    s += super
+    s
+  end
+end
