+++ Library/Homebrew/requirements/arch_requirement.rb
@@ -1,0 +1,21 @@
+require "requirement"
+
+class ArchRequirement < Requirement
+  fatal true
+
+  def initialize(arch)
+    @arch = arch.pop
+    super
+  end
+
+  satisfy(build_env: false) do
+    case @arch
+    when :x86_64 then MacOS.prefer_64_bit?
+    when :intel, :ppc then Hardware::CPU.type == @arch
+    end
+  end
+
+  def message
+    "This formula requires an #{@arch} architecture."
+  end
+end
