+++ Library/Homebrew/test/dependency_collector_spec.rb
@@ -87,7 +87,7 @@ describe DependencyCollector do
     it "creates a resource dependency from a CVS URL" do
       resource = Resource.new
       resource.url(":pserver:anonymous:@example.com:/cvsroot/foo/bar", using: :cvs)
-      expect(subject.add(resource)).to be_an_instance_of(CVSRequirement)
+      expect(subject.add(resource)).to eq(Dependency.new("cvs", [:build]))
     end
 
     it "creates a resource dependency from a Subversion URL" do
